[Code] Refactor the search bar in the source view page to use the same searchbar in other pages
