+++ x-pack/plugins/code/public/components/diff_page/diff.tsx
@@ -14,7 +14,7 @@ import { CommitDiff, FileDiff } from '../../../common/git_diff';
 import { SearchScope } from '../../../model';
 import { changeSearchScope } from '../../actions';
 import { RootState } from '../../reducers';
-import { SearchBar } from '../search_bar';
+// import { SearchBar } from '../search_bar';
 import { DiffEditor } from './diff_editor';
 
 const COMMIT_ID_LENGTH = 16;
@@ -191,11 +191,11 @@ export class DiffPage extends React.Component<Props> {
     ));
     return (
       <div className="diff">
-        <SearchBar
+        {/* <SearchBar
           repoScope={this.props.repoScope}
           query={this.props.query}
           onSearchScopeChanged={this.props.onSearchScopeChanged}
-        />
+        /> */}
         {topBar}
         <Container>
           <EuiText>{commit.commit.message}</EuiText>
